+++ VultisigApp/VultisigApp/Features/VaultCommon/Views/Banners/BannersCarousel.swift
@@ -27,7 +27,7 @@ struct BannersCarousel<Banner: CarouselBannerType>: View {
     }
     
     var horizontalPadding: CGFloat {
-        (availableWidth - bannerWidth) / 2
+        max((availableWidth - bannerWidth) / 2, BannerLayoutProperties.minimumPadding)
     }
     
     private let bannerHeight: CGFloat = 128
@@ -198,71 +198,6 @@ struct BannersCarousel<Banner: CarouselBannerType>: View {
     private func stopTimer() {
         timer?.invalidate()
         timer = nil
-    }
-}
-
-private struct VaultBannerCarouselIndicators: View {
-    @Binding var currentIndex: Int
-    @Binding var bannersCount: Int
-    
-    var body: some View {
-        HStack(spacing: 4) {
-            ForEach(0..<bannersCount, id: \.self) { index in
-                VaultBannerCarouselIndicator(
-                    indicatorIndex: index,
-                    currentIndex: $currentIndex,
-                    bannersCount: bannersCount
-                )
-            }
-        }
-    }
-}
-
-private struct VaultBannerCarouselIndicator: View {
-    let indicatorIndex: Int
-    @Binding var currentIndex: Int
-    let bannersCount: Int
-    
-    @State var isActive: Bool = false
-    @State var progress: CGFloat = 0
-    
-    let capsuleWidth: CGFloat = 20
-    let capsuleHeight: CGFloat = BannerLayoutProperties.indicatorsHeight
-    let animation: Animation = .interpolatingSpring(mass: 1, stiffness: 100, damping: 15)
-    
-    var body: some View {
-        Capsule()
-            .fill(Theme.colors.bgTertiary)
-            .frame(width: isActive ? capsuleWidth : capsuleHeight, height: capsuleHeight)
-            .padding(.horizontal, 0.1)
-            .overlay(isActive ? overlayView : nil, alignment: .leading)
-            .onLoad {
-                updateIsActive()
-            }
-            .onChange(of: currentIndex) {
-                updateIsActive()
-            }
-    }
-    
-    var overlayView: some View {
-        Capsule()
-            .fill(Theme.colors.textLight)
-            .frame(width: progress, height: capsuleHeight + 0.1)
-            .offset(x: -1)
-            .onAppear {
-                withAnimation(.linear(duration: 2.5).delay(0.5)) {
-                    progress = capsuleWidth + 1
-                }
-            }
-    }
-    
-    func updateIsActive() {
-        withAnimation(animation) {
-            isActive = currentIndex == indicatorIndex
-            if !isActive {
-                progress = 0
-            }
-        }
     }
 }
